Allow game to quit gracefully (no exception)
Before, `pygame.error` was being raised, with the message
"video system not initialized" because the game loop would
never end (`game_exit` was never set to `True`). Now, it *is*
set and a check is made to ensure that the main loop ends
*before* any other code past the event loop does.

This is likely to cause a decrease in performance due to the
additional check for `game_exit` in every iteration of the
main loop.