Set a fixed seed (1) to enable deterministic tests
This relies on the game only using the seed it was provided
to generate random levels, enemies, etc.

This helps to ensure that any performance tests are done such
that they are fair tests, i.e., only 1 variable changes between
each test in order to be able to determine that the change
in said variable was the cause of the improvement or regression.